Combine SpecificationSection with ResourcesSection
